requirements.txt
# Core dependencies
google-adk
a2a
agent-starter-pack
pydantic
fastapi
fastapi-utilities
openmemory


mcp
telethon
telegram
python-json-logger
nest-asyncio
python-telegram-bot
aiosqlite

# Google APIs
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# MCP Servers 
git+https://github.com/jacepark12/ticktick-mcp.git

# Database and async support
sqlalchemy
asyncpg
psycopg2-binary
alembic

# Authentication and security
cryptography
python-jose[cryptography]
passlib[bcrypt]

# Utilities
python-dotenv
uvicorn
requests
aiofiles
watchdog

# Testing
pytest
pytest-asyncio
httpx